+++ CHANGELOG.markdown
@@ -1,7 +1,11 @@
 CHANGELOG
 =========
 
-* 1.2.4-DEV (???)
+* 1.2.5-DEV (???)
+
+  * (nothing)
+
+* 1.2.4 (2014-09-28)
 
   * [#52] fix document-template support
   * [#64] fix dependencies versions in composer.json